Split up TypeMapper into several partial Classes
